[AMD] Various build fixes from 25.02 -> 25.04
Most build issues were implicit narrowing from int -> unsigned
int, and automatic type deduction failing. These are fixed
by changing size types to size_t or uint32/64t types.

It is ok to static_cast nnz values since the size should always
be equal to or greater than zero. A negative value would be an
error regardless.

+++ cpp/src/sparse/cluster/detail/spectral.cuh
@@ -62,7 +62,7 @@ void fit_embedding(raft::resources const& handle,
   value_type* vs = dst_vals.data();
 
   raft::spectral::matrix::sparse_matrix_t<index_type, value_type> const r_csr_m{
-    handle, ro, ci, vs, n, nnz};
+    handle, ro, ci, vs, n, static_cast<size_t>(nnz)};
 
   index_type neigvs       = n_components + 1;
   index_type maxiter      = 4000;  // default reset value (when set to 0);
